requirements.txt
djangorestframework
mysqlclient
djangorestframework-simplejwt
timedelta
pandas